# app/src/main/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TP1 - Teléfono descompuesto",
    "description": "Title",
    "version": "1.0.0"
  },
  "paths": {
    "/register-node": {
      "post": {
        "operationId": "registerNode",
        "description": "Registra un nuevo nodo",
        "parameters": [
          {
            "name": "host",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "port",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo bien",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/registerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/relay": {
      "post": {
        "operationId": "relayMessage",
        "description": "Firma un mensaje y lo manda al siguiente",
        "x-content-type": "multipart/form-data",
        "requestBody": {
          "description": "El mensaje y las firmas anteriores",
          "content": {
            "multipart/form-data": {
              "encoding": {
                "message": {
                  "contentType": "*/*"
                },
                "signatures": {
                  "contentType": "application/json"
                }
              },
              "schema": {
                "type": "object",
                "required": [
                  "message",
                  "signatures"
                ],
                "properties": {
                  "message": {
                    "$ref": "#/components/schemas/base64String"
                  },
                  "signatures": {
                    "$ref": "#/components/schemas/signatures"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "mensaje recibido y reenviado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/signature"
                }
              }
            }
          }
        }
      }
    },
    "/play": {
      "post": {
        "operationId": "sendMessage",
        "description": "Comienza el juego!",
        "requestBody": {
          "description": "El mensaje a enviar por la red telefónica",
          "required": true,
          "content": {
            "*/*": {
              "schema": {
                "type": "string",
                "format": "base64"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "La red telefónica funcionó bien!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/playResponse"
                }
              }
            }
          },
          "503": {
            "description": "La red telefónica falló",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/playResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "base64String": {
        "type": "string",
        "format": "base64"
      },
      "registerRequest": {
        "type": "object",
        "required": [
          "host",
          "port",
          "name"
        ],
        "properties": {
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "registerResponse": {
        "type": "object",
        "required": [
          "nextHost",
          "nextPort",
          "uuid",
          "hash"
        ],
        "properties": {
          "nextHost": {
            "type": "string"
          },
          "nextPort": {
            "type": "integer"
          },
          "uuid": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          }
        }
      },
      "signature": {
        "type": "object",
        "required": [
          "name",
          "hash"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "El nombre del nodo que firmo"
          },
          "contentType": {
            "type": "string",
            "description": "El hash de la firma del nodo"
          },
          "contentLength": {
            "type": "integer",
            "description": "La longitud del contenido"
          },
          "hash": {
            "type": "string",
            "description": "El hash del contenido calculado por el nodo"
          }
        }
      },
      "signatures": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/signature"
            }
          }
        }
      },
      "playResponse": {
        "type": "object",
        "required": [
          "contentResult",
          "originalContentType",
          "originalLength",
          "originalHash",
          "receivedContentType",
          "receivedLength",
          "receivedHash",
          "signatures"
        ],
        "properties": {
          "contentResult": {
            "type": "string",
            "description": "el resultado del contenido"
          },
          "originalContentType": {
            "type": "string",
            "description": "el tipo de contenido que enviaste"
          },
          "originalLength": {
            "type": "integer",
            "description": "la longitud original que enviaste"
          },
          "originalHash": {
            "type": "string",
            "description": "el hash del mensaje original que enviaste"
          },
          "receivedContentType": {
            "type": "string",
            "description": "el tipo de contenido que volvió"
          },
          "receivedLength": {
            "type": "integer",
            "description": "la longitud que volvió"
          },
          "receivedHash": {
            "type": "string",
            "description": "el hash del mensaje que volvió"
          },
          "signatures": {
            "$ref": "#/components/schemas/signatures"
          }
        }
      }
    }
  }
}